Move all DB seed data into seed/ package; init_env.py now bootstraps everything required
