+++ tests/conftest.py
@@ -10,8 +10,9 @@ import tempfile
 
 import pytest
 
-from app import create_app, DEFAULT_CATEGORIES, DEFAULT_LOCATIONS
+from app import create_app
 from app.models import db as _db, Category, State, City, User
+from seed.initial_data import DEFAULT_CATEGORIES, DEFAULT_LOCATIONS
 
 
 @pytest.fixture()
